status and state codes are back (#6)
create a _display named attribute that is read-only and described as 'inSupportOf" the contrainedList type integed values attributes.  Removed Annotations from schema to avoid abuse. added an indicator for the correlated value for the situation where a constrainedList has values that are dependent on the values of a different attribute (status and state) Bumped the version that causes all IDs to change