new cdm schema document proposal and object model for tools
entities: The 'correct' (based on dependency hierarchy) arrangement of entities / extensions / folders of the CRM entities. Addition of the entities for the ‘new’ marketing application
entity document: new document format that allows for easy expression of semantic meanings for entities and attributes.
Supporting tools: the CDM object model that interprets / produces the JSON docs for use with any future CDM tooling.