updates for new dplx support
changed the way that attributeGroups get identified in entity. went to one trait with a table of entries rather than creating new traits for each path. that meant I needed to merge constant entity rows when resolving traits. this needed to be done anyway for cases like when extension entities have duplicate attributes that are only put there to  add new status values to an option set