fix some name issues. portals. browser search. OM cleanup. new dplx
portals entities. make a dependable and regular set of attribute and entity names from xrm. major cleanup of OM. removed the string class and the attribute promise idea and replaced with regular referenece objects that know they hold a string only. this is in prep for the next step of OM work that fixes the problems with references to base classes