refactor continues. test spew.
import sequence number added to standard set. added some test code to spew all things learned in resolving entities. remove the code path for copy on write of trait sets because with the new way of making a copy on reference, it never gets used. change to  argument setting to only use override values if a values was explicitly set.