step 1 of the new 'resolved entity' system
switching to using a resolved entity as a source of 'result of resolution' instead of the getResolvedXXX api calls. So, all context discovered during resolution is now captured in the entity document as a AttributeContext hierarchy. Also, much cleaner system for representing polymorphic relationships to multiple entities. (will become the 'pick one attribute from the embedded entity' relationship in the next phase)