finish flexibleRelationships. getting ready for .8
re-worked all of the code for how the trait appliers work. it was just a disaster and it wasn't working for my test of flexible relationships (making contact and account actually link to address entity and bring in an array expansion). I updated all schema docs that use polymorphic relationships to use the new system.