performance improvement for OM resolution
solved the potential n*n (entities) performance problem and others by use of very careful caching. key insight is that many symbols are only defined in one document and can only have one resolution independent of order of imports. For others, (like when traits have parameters that change per reference or symbols that get re-defined, the trick was to figure out a unique way to identify the context for caching that could be shared with similar but not identical contexts because the cached version was context independent. To do this, needed to index every declaration of a symbol in the entire corpus and then, given a reference in the context of some set of imports figure out exactly which definition would get used (also depends on any objects referenced from the object being referenced, those refs come from some set of documents) Cache resolved references and definition and traits and attributes based on the list of source documents found in the import of the given context. Many, many contexts that have different import lists will in fact have the identical set of 'definition' documents for a reference. so, avoids a ton of code.